# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Echobot",
    "description": "AI驱动的数字人动画平台 API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/stories": {
      "get": {
        "tags": [
          "故事"
        ],
        "summary": "获取故事列表",
        "description": "获取已发布的故事列表\n\n支持分页、分类筛选和关键词搜索。\n普通用户只能查看已发布状态的故事。\n\n**请求参数:**\n- **page**: 页码，从1开始\n- **page_size**: 每页数量，1-100\n- **category_id**: 分类ID（可选），筛选特定分类下的故事\n- **search**: 搜索关键词（可选），匹配故事标题\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"items\": [\n            {\n                \"id\": \"694534f1c32ffbd6151c18a2\",\n                \"title\": \"三只小猪\",\n                \"title_en\": \"The Three Little Pigs\",\n                \"category_id\": \"694534f1c32ffbd6151c18a1\",\n                \"thumbnail_url\": \"/uploads/thumbnails/story123.jpg\",\n                \"duration\": 180,\n                \"is_published\": true,\n                \"view_count\": 1234,\n                \"created_at\": \"2025-01-01T12:00:00\"\n            }\n        ],\n        \"total\": 50,\n        \"page\": 1,\n        \"page_size\": 20\n    }\n}\n```",
        "operationId": "list_stories_api_v1_stories_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码，从1开始",
              "default": 1,
              "title": "Page"
            },
            "description": "页码，从1开始"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量，最大100",
              "default": 20,
              "title": "Page Size"
            },
            "description": "每页数量，最大100"
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "分类ID，用于筛选特定分类",
              "title": "Category Id"
            },
            "description": "分类ID，用于筛选特定分类"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "搜索关键词，匹配标题",
              "title": "Search"
            },
            "description": "搜索关键词，匹配标题"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/stories/{story_id}": {
      "get": {
        "tags": [
          "故事"
        ],
        "summary": "获取故事详情",
        "description": "获取单个故事的完整详情\n\n返回故事的所有信息，包括视频URL、字幕数据等。\n\n**路径参数:**\n- **story_id**: 故事唯一标识ID\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"id\": \"694534f1c32ffbd6151c18a2\",\n        \"title\": \"三只小猪\",\n        \"title_en\": \"The Three Little Pigs\",\n        \"category_id\": \"694534f1c32ffbd6151c18a1\",\n        \"category_name\": \"童话故事\",\n        \"description\": \"三只小猪建房子的经典童话故事\",\n        \"thumbnail_url\": \"/uploads/thumbnails/story123.jpg\",\n        \"video_url\": \"/uploads/videos/story123.mp4\",\n        \"audio_url\": \"/uploads/audio/story123.mp3\",\n        \"duration\": 180,\n        \"is_published\": true,\n        \"view_count\": 1234,\n        \"subtitles\": [\n            {\"start\": 0.0, \"end\": 3.5, \"text\": \"从前有三只小猪...\"}\n        ],\n        \"created_at\": \"2025-01-01T12:00:00\",\n        \"updated_at\": \"2025-01-01T12:00:00\"\n    }\n}\n```",
        "operationId": "get_story_api_v1_stories__story_id__get",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Story Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/audiobook/stories": {
      "get": {
        "tags": [
          "管理后台"
        ],
        "summary": "获取有声书故事列表（管理端）",
        "description": "获取有声书故事列表（管理端）",
        "operationId": "list_audiobook_stories_api_v1_admin_audiobook_stories_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "is_published",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Published"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "管理后台"
        ],
        "summary": "创建有声书故事",
        "description": "创建有声书故事",
        "operationId": "create_audiobook_story_api_v1_admin_audiobook_stories_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_audiobook_story_api_v1_admin_audiobook_stories_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/audiobook/stories/{story_id}": {
      "get": {
        "tags": [
          "管理后台"
        ],
        "summary": "获取有声书故事详情",
        "description": "获取有声书故事详情",
        "operationId": "get_audiobook_story_api_v1_admin_audiobook_stories__story_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Story Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "管理后台"
        ],
        "summary": "更新有声书故事",
        "description": "更新有声书故事",
        "operationId": "update_audiobook_story_api_v1_admin_audiobook_stories__story_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Story Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_audiobook_story_api_v1_admin_audiobook_stories__story_id__put"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "管理后台"
        ],
        "summary": "删除有声书故事",
        "description": "删除有声书故事",
        "operationId": "delete_audiobook_story_api_v1_admin_audiobook_stories__story_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Story Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/audiobook/jobs": {
      "get": {
        "tags": [
          "管理后台"
        ],
        "summary": "获取有声书任务列表（管理端）",
        "description": "获取有声书生成任务列表（管理端）",
        "operationId": "list_audiobook_jobs_api_v1_admin_audiobook_jobs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/audiobook/stories": {
      "get": {
        "tags": [
          "有声书"
        ],
        "summary": "获取有声书故事列表",
        "description": "获取有声书故事列表\n\n返回平台上所有可用的有声书故事模板。\n支持按语言、分类、年龄段进行筛选。\n\n**请求参数:**\n- **page**: 页码，从1开始\n- **page_size**: 每页数量，1-100\n- **language**: 语言筛选（可选）\n- **category**: 分类筛选（可选）\n- **age_group**: 年龄段筛选（可选）\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"items\": [\n            {\n                \"id\": \"695342ec383483bf26139d89\",\n                \"title\": \"三只小猪\",\n                \"title_en\": \"The Three Little Pigs\",\n                \"content\": \"Once upon a time...\",\n                \"language\": \"en\",\n                \"category\": \"fairy_tale\",\n                \"age_group\": \"5-8\",\n                \"estimated_duration\": 180,\n                \"thumbnail_url\": \"/uploads/audiobook/cover.jpg\",\n                \"is_published\": true\n            }\n        ],\n        \"total\": 10,\n        \"page\": 1,\n        \"page_size\": 20\n    }\n}\n```",
        "operationId": "list_stories_api_v1_audiobook_stories_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码，从1开始",
              "default": 1,
              "title": "Page"
            },
            "description": "页码，从1开始"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量，最大100",
              "default": 20,
              "title": "Page Size"
            },
            "description": "每页数量，最大100"
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "语言筛选：en(英语) | zh(中文)",
              "title": "Language"
            },
            "description": "语言筛选：en(英语) | zh(中文)"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "分类筛选：fairy_tale | fable | adventure | bedtime",
              "title": "Category"
            },
            "description": "分类筛选：fairy_tale | fable | adventure | bedtime"
          },
          {
            "name": "age_group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "年龄段筛选：3-5 | 5-8 | 8-12",
              "title": "Age Group"
            },
            "description": "年龄段筛选：3-5 | 5-8 | 8-12"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audiobook/stories/{story_id}": {
      "get": {
        "tags": [
          "有声书"
        ],
        "summary": "获取有声书故事详情",
        "description": "获取有声书故事详情\n\n返回故事的完整内容，包括用于 TTS 合成的文本。\n\n**路径参数:**\n- **story_id**: 故事唯一标识ID\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"id\": \"695342ec383483bf26139d89\",\n        \"title\": \"三只小猪\",\n        \"title_en\": \"The Three Little Pigs\",\n        \"content\": \"Once upon a time, there were three little pigs who lived with their mother...\",\n        \"language\": \"en\",\n        \"category\": \"fairy_tale\",\n        \"age_group\": \"5-8\",\n        \"estimated_duration\": 180,\n        \"thumbnail_url\": \"/uploads/audiobook/cover.jpg\",\n        \"background_music_url\": \"/uploads/audiobook/bgm.mp3\",\n        \"is_published\": true,\n        \"created_at\": \"2025-01-01T12:00:00\",\n        \"updated_at\": \"2025-01-01T12:00:00\"\n    }\n}\n```",
        "operationId": "get_story_api_v1_audiobook_stories__story_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Story Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/audiobook/jobs/{job_id}": {
      "get": {
        "tags": [
          "有声书"
        ],
        "summary": "获取有声书任务状态",
        "description": "获取有声书任务状态\n\n轮询此接口获取任务的处理进度和结果。\n建议轮询间隔：2-3秒。\n\n**路径参数:**\n- **job_id**: 任务ID（从 POST /jobs 返回）\n\n**任务状态说明:**\n- **pending**: 等待处理\n- **processing**: 处理中\n- **completed**: 完成，audio_url 包含生成的音频\n- **failed**: 失败，error 包含错误信息\n\n**处理步骤说明:**\n- **init**: 初始化任务\n- **tts**: 语音合成中\n- **mixing**: 音频混合中（添加背景音乐）\n- **completed**: 处理完成\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"id\": \"695345efc32ffbd6151c18a9\",\n        \"user_id\": \"6948bda91fd873cf81f5addd\",\n        \"story_id\": \"695342ec383483bf26139d89\",\n        \"voice_profile_id\": \"694534f1c32ffbd6151c18a3\",\n        \"status\": \"completed\",\n        \"progress\": 100,\n        \"current_step\": \"completed\",\n        \"audio_url\": \"/uploads/audiobook/user123/output.mp3\",\n        \"duration\": 185,\n        \"story_title\": \"三只小猪\",\n        \"voice_name\": \"爸爸的声音\",\n        \"created_at\": \"2025-01-01T12:00:00\",\n        \"completed_at\": \"2025-01-01T12:03:00\",\n        \"error\": null\n    }\n}\n```",
        "operationId": "get_job_api_v1_audiobook_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_audiobook_story_api_v1_admin_audiobook_stories_post": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "title_en": {
            "type": "string",
            "title": "Title En"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "en"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "fairy_tale"
          },
          "age_group": {
            "type": "string",
            "title": "Age Group",
            "default": "5-8"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "background_music_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Background Music Url"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published",
            "default": true
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "title",
          "title_en",
          "content"
        ],
        "title": "Body_create_audiobook_story_api_v1_admin_audiobook_stories_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}